asignacion de permisos y roles a los usuarios

+++ app/Http/Controllers/BrandController.php
@@ -5,12 +5,14 @@ use App\Http\Requests\Brand\StoreRequest;
 use App\Http\Requests\Brand\UpdateRequest;
 use App\Brand;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class BrandController extends Controller
 {
   
     public function index(Request $request)
     {
+        abort_if(Gate::denies('brand_index'), 403);
         $nombre = $request->get('buscar-nombre');
        
         $brands = Brand::nombres($nombre)->paginate(5);
@@ -22,6 +24,7 @@ class BrandController extends Controller
 
     public function create()
     {
+        abort_if(Gate::denies('brand_create'), 403);
         $brand = new Brand();
         return view('admin.brand.create', compact('brand'));
 
@@ -45,6 +48,7 @@ class BrandController extends Controller
    
     public function edit(Brand $brand)
     {
+        abort_if(Gate::denies('brand_edit'), 403);
         $brands = Brand::get();
         return view('admin.brand.edit', compact('brand'));
     }
@@ -60,6 +64,7 @@ class BrandController extends Controller
     
     public function destroy(Brand $brand)
     {
+        abort_if(Gate::denies('brand_destroy'), 403);
         'alert'()->success('Exito','Se ha eliminado el registro.');
         $brand->delete();
         return redirect()->route('brands.index');
