+++ app/Http/Controllers/DealerController.php
@@ -4,12 +4,14 @@ namespace App\Http\Controllers;
 
 use App\Dealer;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class DealerController extends Controller
 {
 
     public function index(Request $request)
     {
+        abort_if(Gate::denies('dealer_index'), 403);
         $nombre = $request->get('buscar-nombre');
         
         $nit = $request->get('buscar-nit');
@@ -23,6 +25,7 @@ class DealerController extends Controller
  
     public function create()
     {
+        abort_if(Gate::denies('dealer_create'), 403);
         $dealer = new Dealer();
         return view('admin.dealer.create', compact('dealer'));
     }
@@ -37,6 +40,7 @@ class DealerController extends Controller
     }
     public function show($id)
     {
+        abort_if(Gate::denies('dealer_show'), 403);
         $dealer = Dealer::find($id);
         return view('admin.dealer.show', compact('dealer'));
         
@@ -44,6 +48,7 @@ class DealerController extends Controller
 
     public function edit($id)
     {
+        abort_if(Gate::denies('dealer_edit'), 403);
         $dealer = Dealer::find($id);
 
         return view('admin.dealer.edit', compact('dealer'));
@@ -63,6 +68,7 @@ class DealerController extends Controller
 
     public function destroy($id)
     {
+        abort_if(Gate::denies('dealer_destroy'), 403);
         'alert'()->success('Exito','Se ha eliminado el registro.');
         $dealer = Dealer::find($id)->delete();
         return redirect()->route('dealers.index');
