+++ app/Http/Controllers/PersonaController.php
@@ -4,12 +4,14 @@ namespace App\Http\Controllers;
 
 use App\Persona;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PersonaController extends Controller
 {
 
     public function index(Request $request)
     {
+        abort_if(Gate::denies('persona_index'), 403);
         $nombre = $request->get('buscar-nombre');
         
         $cedula = $request->get('buscar-cedula');
@@ -21,6 +23,7 @@ class PersonaController extends Controller
     
     public function create()
     {
+        abort_if(Gate::denies('persona_create'), 403);
         $persona = new Persona();
         return view('admin.persona.create', compact('persona'));
     }
@@ -38,6 +41,7 @@ class PersonaController extends Controller
 
     public function show($id)
     {
+        abort_if(Gate::denies('persona_show'), 403);
         $persona = Persona::find($id);
 
         return view('admin.persona.show', compact('persona'));
@@ -46,6 +50,7 @@ class PersonaController extends Controller
 
     public function edit($id)
     {
+        abort_if(Gate::denies('persona_edit'), 403);
         $persona = Persona::find($id);
 
         return view('admin.persona.edit', compact('persona'));
@@ -64,6 +69,7 @@ class PersonaController extends Controller
 
     public function destroy($id)
     {
+        abort_if(Gate::denies('persona_destroy'), 403);
         'alert'()->success('Exito','Se ha eliminado el registro.');
         $persona = Persona::find($id)->delete();
         return redirect()->route('personas.index')
