+++ app/Http/Controllers/RoleController.php
@@ -12,7 +12,7 @@ class RoleController extends Controller
    
     public function index()
     {
-        // abort_if(Gate::denies('role_index'), 403);
+        abort_if(Gate::denies('role_index'), 403);
         // $permissions = Permission::paginate(10);
         $roles = Role::paginate(10);
 
@@ -22,7 +22,7 @@ class RoleController extends Controller
    
     public function create()
     {
-        // abort_if(Gate::denies('role_create'), 403);
+        abort_if(Gate::denies('role_create'), 403);
         
         
         $permissions = Permission::all()->pluck('name', 'id');
@@ -44,7 +44,7 @@ class RoleController extends Controller
    
     public function show(Role $role)
     {
-        // abort_if(Gate::denies('role_show'), 403);
+        abort_if(Gate::denies('role_show'), 403);
 
         $role->load('permissions');
         return view('admin.roles.show', compact('role'));
@@ -53,7 +53,7 @@ class RoleController extends Controller
   
     public function edit(Role $role)
     {
-        // abort_if(Gate::denies('role_edit'), 403);
+        abort_if(Gate::denies('role_edit'), 403);
       
         $permissions = Permission::all()->pluck('name', 'id');
         $role->load('permissions');
@@ -75,7 +75,7 @@ class RoleController extends Controller
    
     public function destroy(Role $role)
     {
-        // abort_if(Gate::denies('role_delete'), 403);
+        abort_if(Gate::denies('role_delete'), 403);
 
         $role->delete();
